third_party/blink/web_tests/platform/webview/external/wpt/pointerevents/pointerevent_after_target_appended_touch-expected.txt

This is a testharness.js-based test.
[FAIL] pointer events from touch received before/after child attached at pointerdown
  assert_equals: events received expected "pointerover@parent,pointerenter@parent,pointerdown@parent,(child-attached),pointerout@parent,pointerover@child,pointerenter@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent,pointerover@child,pointerenter@child,pointerenter@parent,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@parent,pointerenter@parent,pointerdown@parent,(child-attached),pointerup@parent,pointerout@parent,pointerleave@parent,pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
[FAIL] pointer events from touch received before/after child attached at pointerup
  assert_equals: events received expected "pointerover@parent,pointerenter@parent,pointerdown@parent,pointerup@parent,(child-attached),pointerout@parent,pointerover@child,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@parent,pointerenter@parent,pointerdown@parent,pointerup@parent,(child-attached),pointerout@parent,pointerleave@parent,pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
[FAIL] pointer events from touch received before/after child moved at pointerdown
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-moved),pointerover@child,pointerenter@child,pointerup@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-moved),pointerover@child,pointerenter@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent,pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
[FAIL] pointer events from touch received before/after child moved at pointerup
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-moved),pointerover@child,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-moved),pointerleave@parent,pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
[FAIL] mouse events from touch received before/after child attached at mousedown
  assert_equals: events received expected "mouseover@parent,mouseenter@parent,mousedown@parent,(child-attached),mouseout@parent,mouseover@child,mouseenter@child,mouseup@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent" but got "mouseover@parent,mouseenter@parent,mousedown@parent,(child-attached),mouseout@parent,mouseover@child,mouseenter@child,mouseup@child,mouseout@child,mouseleave@child,mouseover@parent,mouseout@parent,mouseover@child,mouseenter@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent"
Harness: the test ran to completion.
